Added core code (#1)
* TxBuilder, BlockBuilder

* Added CRC256

CRC256 is a 256 bit cyclic redundancy check based on the security of the hardware implemented CRC32. By executing the CRC32 many times in parallel, speeds of over 8 bytes per CPU cycle (on x86) are possible. ARMv7 and ARMv8 CPUs are even stronger.

This allows the usage of a hash faster than Blake2 for validation of the block hash. CRC256 can be extended to CRC512 easily.

* Added CRC512

* Completeted blocks+transactions

* Moved up

* Added config

* Fixed BlockSize

* Diff to 64bit integer

* Added validation code

* Added getHeader function

* Added CPU miner

* Extended config (QUICK_VERIFY)

* Added validation header

* Added header guards